tools/mcp_robot/requirements.txt
fastmcp
pyntcore
